requirements.txt
# Core Framework
Django>=4.2.11,<5.0
dj-database-url==2.1.0
whitenoise[brotli]==6.8.2
gunicorn==22.0.0
django-allauth==0.61.1

# AI Orchestration & Structured Output
# Pinning instructor to 1.14.x as it's the stable 2026 baseline
instructor==1.14.5
litellm==1.63.3
pydantic>=2.6.3,<3.0.0

# Critical Fix: Pin Mistral SDK to < 2.0.0 to resolve the Instructor import bug
mistralai>=1.0.0,<2.0.0

# Storage & Utilities
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
psycopg2-binary==2.9.9
requests>=2.32.3
asgiref>=3.8.1
python-dateutil>=2.9.0